[REF] supply_chain_report_jsi: add update button
The supply chain report is heavy on the database if the fields are store
with a compute. To fix that we previously unstored the fields. The issue
with that approach is that the customer can no longer sort his data
using the fields.

Our solution for the moment is to store the field but without a compute
method having depends. We just add a button which recomputes the data,
(+-1min15sec of recompute atm). The report is consulted daily so this is
acceptable.